Move PaparazziComposeOwner into its own file
